Attempt to optimize math operations
Since, for each mandelbrot iteration, several variables are added/subtracted/compared, do the normalizing all in one function and in-place where possible (as opposed to creating new InfNum objects).  In theory this would result in fewer function calls and prevent normalizing the same value multiple times.

In my testing though this has essentially zero impact, positive or negative, on the running time when generating an image.  More work needs to be done with this.  For simplicity this code may be removed in the future.