Truncate scale/center to resolve slowdowns
As zoom and pan are used, the scale and centerX/centerY parameters quickly accumulate many digits.  After only a few pans and zooms, the calculations are slowed down to half speed, even when floating point is used for Mandelbrot set.  The slowdown can be resolved by refreshing the page (by using the truncated parameters in the URL).  By truncating the scale and centerX/centerY to 20 significant digits before every run, this slowdown is avoided without having to refresh the page.