Workaround since presets may require more precision
If, after starting at the default Mandelbrot set view, the user presses the "1" key to view the first preset, the window would not be centered properly on the intended location.  This is because the default Mandelbrot set has a small scale value and thus sets the global precision to 12 digits.  The first preset looks like it needs at least 16 digits of precision to center the view as intended.

With this change, a preset can declare some level of precision, which will be set before the preset is applied to the page URL.