WIP: refactor to web workers
Some changes I've been making lately have been resulting in sluggish performance and "out of memory" issues on my M1 Mac mini (late 2020) when using Firefox.  Safari on the M1 Mac seems to run ok.  But on Firefox, very quickly after loading the page and panning/zooming around a little, I encounter the sluggishness and out of memory error.  Even on non-window plots.  These problems are not occurring on my Intel i5 Ubuntu machine in Firefox nor in Chrome.

I just updated Firefox to 94.0.2 and perhaps, maybe, the problem was with Firefox, not my code, and was fixed.  There was what seemed to be a graphics-related memory leak that was addressed by this Firefox update, so that could explain it.

As far as this commit goes, I'm ripping things apart and moving things around.  I have most of the structure of the web workers in place.  The workers will need to load a bunch of code (the InfNum stuff, at least) so that will have to be moved into a separate file or function.  The main/UI thread still does a synchronous 64-wide pass, and a new version of that seems to be working.  The worker stuff is not active in this commit -- just the synchronous 64-wide pass.

This first attempt at web workers has no caching, for now.  The cache will have to be moved into the main worker, but that can happen later once the rest of it is working.