Refactor cache scan objects to use Map
Since we want to store chunk index integers as keys, using Objects doesn't work so well because all Object keys are converted to Strings.  I think due to this, the subworker was checking for cached indices and never finding any because it was looking for int values and was being passed Object.keys() values, which were actually Strings.  Anyway.  Using Map also means we can use the .size attribute which saves us from having to count each cached point individually.