Use binary search for cached points
The subworker is given an array of points indices that are cached.  Any point along the chunk that is cached can be skipped.  Using a binary search should make this more efficient.  The main worker thread sorts the array of cached indices ahead of time, in a different thread, so sorting the array theoretically has no impact on performance.

With this binary search in place:

where no points are initially cached (each pass has 25% of its points cached from the previous pass):
- a 1.7 megapixel image at n=5000 is rendered around 1% faster
- a 3.8 megapixel image at n=5000 is rendered around 2% faster
- a 1.7 megapixel image at n=40 is rendered around 7.6% faster

where all points are initially cached:
- a 1.7 megapixel image at n=5000 is rendered 0.8% faster
- a 3.8 megapixel image at n=5000 is rendered 1.6% faster
- a 1.7 megapixel image at n=40 is rendered in the same time