Add function to use polygon method to find period.
This method is explained here: http://www.mrob.com/pub/muency/period.html

I believe then every point inside the polygon (box) will have the found period, but I'm not confident of this.

As to how this is useful, I'm not sure.  For Zhuoran's perturbation algorithm, we can re-base at the end of a reference orbit, so the reference orbit doesn't need to continue past its period.  This could speed up reference orbit calculation if not for the fact that in order to find the period, 4 points must be iterated with high precision...