Simplify non-worker mode by removing points cache
Since we have the pixel cache for quickly changing colors, we can make computation faster and simpler by removing the concept of a points cache.  Since the points cache really only helps when panning, and most exploration would be with zooming, not panning, the points cache wasn't used much anyway.  Also, the logic for skipping every other pixel in every other row is faster to use than relying on the cache to find computed pixels from a previously-completed pass.