Add "mod" gradients that take mod of pixel iterations
This allows much smaller, faster gradients to be used for high-iteration locations.  Previously, with percentage-based gradients, some high-iteration locations required hundreds of color stops for decent-looking renders.

The "mod" gradients are incompatible with sequence plots where each line is colored according to its percentage "distance" down the overall line, so lots of checking has to be done to ensure the proper gradient type is applied to a plot.