Polyfill for structuredClone() for Chrome/Safari
Apparently Firefox is ahead of the curve with structuredClone().  Luckily, it looks like Chrome will support it as of version 98 which will be released soon.  Also luckily, structuredClone() isn't used for any terribly complex objects and I was able to make a quick and dirty polyfill that works well and quickly enough.