Restrict magnification/scale in some places
- When using zoom box, restrict computed scale to 5 significant digits.
- When doing any magnification/scale conversion, restrict to 12 significant digits, not 24

Also, when going to a bounding box -defined location, first calculate the rough scale using the given bounds and the previous precision setting.  Then, re-calculate the scale again using the plot's precision setting for that scale.  The use case for this is when first opening the site and going to the default Mandelbrot set location, the precision will be set rather low.  If the user then enters a set of high-precision bounding box values in the control menu, we don't want to use that original low precision value when computing the scale value from those high-precision bounds.