Keep separate gradients for mod and pct plots
When applying a "-mod" option to the gradient for the Mandelbrot set, then choosing a sequence plot (with X key, or choosing it from the plots listing) the drawn plot would have a broken-looking gradient because "-mod" gradients are not compatible with sequence plots.  Now, when drawing a new plot, if the existing gradient is incompatible, we swap out the gradient.  Since there are only two gradient types, we can just swap whenever the type doesn't match.

This can still break (for example, after choosing the "1" preset (which sets a "-mod" gradient) the user very quickly hits the X key, the URL location wouldn't have updated yet and the "-mod" gradient would then get applied to the primes sequence plot) so a better solution should be used in the future.