Add -mod100 to default Mandelbrot plot, and use plot forcedDefaults
In order to not apply this "-mod100" to the default gradient for
sequence plots (when the plot name and URL version param are included,
but no gradient is included, in the user's entered URL params) use a
different default gradient, without "-mod100" when the plot is a
sequence plot.

Also use the user-specified plot's "forcedDefaults" to override the
default params.  If, for example, the user omits the "n" (iterations)
parameter from the URL, a sensible default for the plot is used.
Previously, the mandelbrot default n=60 would be used, which is not a
good default at all for the "Primes-1-Step-90-turn" plot.