Add stripe average coloring
Escape time and perturbation theory, for floating point only, algorithms
now have an option to use "stripe average coloring."  This is done with
a new computeBoundPointColorStripes() function in plots.js for the
escape time algorithm and with modifications to the
computeBoundPointColorPerturbOrBla() function to optionally perform the
stripe average computation.

This coloring can be used with the window "algo" URL param, with a new
"auto-stripes" algorithm name, or with an explicit "basic-stripes-..."
or "perturb-stripes-..." algorithm name.  Some suggested options have
been added to the "options" dropdown menu in the "algorithm options"
section of the controls menu.