Somewhat working BLA implementation
Added calculation of BLAs.  1-iteration BLAs are calculated for each
reference orbit iteration, then merged in a binary fashion, where the
reference orbit iteration index 1 (the 2nd iteration) has the largest
ultimate BLA.

In the perturbation loop, first check to see if any BLAs are valid for
the reference orbit iteration, and apply the largest one.  If not, run a
regular perturbation step.

This is all working fine, but it's slower than series approximation.
More testing with faster+better data structures is needed.