Re-calc BLAs if window moves, and allow cancellation
Save and resume the looping states of the minibrot/period-finding
functions with context objects.  This unfortunately makes the code much
harder to read, but allows the user to cancel a long-running operation
with the esc key.

Also send more status updates back to the main thread so the progress of
long-running calculations can be displayed to the user.