Add faster infNumTruncateToLen() function
According to the performance tests, it runs about 3x as fast as the
previous version.  This made a noticeable increase in arbitrary
precision processing speed (used for reference orbit calcs, ball
arithmetic, and newton's method).

Also refactored infnum unit tests to only output to the console if a
test fails.